# services/document-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing documents and file storage",
        "title": "Document Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/documents/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get document details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Get document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Document"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update document metadata",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Update document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.DocumentUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document updated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Document"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a document (uploader or admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "documents"
                ],
                "summary": "Delete document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document deleted",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Collection": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "documents": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Document"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "metadata": {
                    "$ref": "#/definitions/models.MetadataSchema"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/models.User"
                },
                "owner_id": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/models.CollectionSettings"
                },
                "slug": {
                    "type": "string"
                },
                "stats": {
                    "description": "Not stored in DB, computed",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.CollectionStats"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.CollectionSettings": {
            "type": "object",
            "properties": {
                "allow_public_submissions": {
                    "type": "boolean"
                },
                "allowed_file_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max_file_size": {
                    "type": "integer"
                },
                "require_approval": {
                    "type": "boolean"
                }
            }
        },
        "models.CollectionStats": {
            "type": "object",
            "properties": {
                "document_count": {
                    "type": "integer"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
        "models.Document": {
            "type": "object",
            "properties": {
                "collection": {
                    "$ref": "#/definitions/models.Collection"
                },
                "collection_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "download_count": {
                    "type": "integer"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_type": {
                    "description": "PDF, HTML, DOCX, etc.",
                    "type": "string"
                },
                "hash": {
                    "description": "SHA-256 for deduplication",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "indexed_at": {
                    "type": "string"
                },
                "is_indexed": {
                    "description": "Processing",
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "metadata": {
                    "description": "Metadata",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.DocumentMetadata"
                        }
                    ]
                },
                "mime_type": {
                    "type": "string"
                },
                "original_filename": {
                    "description": "File information",
                    "type": "string"
                },
                "page_count": {
                    "type": "integer"
                },
                "processing_error": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.DocumentStatus"
                },
                "storage_path": {
                    "description": "S3 key or path",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uploader": {
                    "$ref": "#/definitions/models.User"
                },
                "uploader_id": {
                    "type": "string"
                },
                "versions": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DocumentVersion"
                    }
                },
                "view_count": {
                    "description": "Stats",
                    "type": "integer"
                }
            }
        },
        "models.DocumentMetadata": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "custom_fields": {
                    "type": "object",
                    "additionalProperties": true
                },
                "isbn": {
                    "type": "string"
                },
                "publish_date": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.DocumentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "active",
                "rejected",
                "archived"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessing",
                "StatusActive",
                "StatusRejected",
                "StatusArchived"
            ]
        },
        "models.DocumentVersion": {
            "type": "object",
            "properties": {
                "change_log": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "document_id": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "storage_path": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "version_number": {
                    "type": "integer"
                }
            }
        },
        "models.MetadataSchema": {
            "type": "object",
            "additionalProperties": true
        },
        "models.User": {
            "type": "object",
            "properties": {
                "collections": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Collection"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "documents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Document"
                    }
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/models.UserRole"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.UserRole": {
            "type": "string",
            "enum": [
                "admin",
                "librarian",
                "patron"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleLibrarian",
                "RolePatron"
            ]
        },
        "response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorInfo"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "service.DocumentUpdate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.DocumentMetadata"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}